Downloading and uploading server handle separated by blueprints
