+++ static/windowUpload.js
@@ -16,7 +16,7 @@ var WindowUpload = {
                 value: "Загрузить",
                 name: "files",
                 link: "uploadedFilesList",
-                upload: "/uploadBags",
+                upload: "/load/upload",
                 on: {
                     onUploadComplete: function(){
                         // TODO update data table after uploading complete
